Rearrange namespaces
