+++ StreamRegex.Benchmarks/BufferSizeBenchmarks.cs
@@ -1,6 +1,7 @@
 ﻿using System.Text.RegularExpressions;
 using BenchmarkDotNet.Attributes;
-using StreamRegex.Extensions;
+using StreamRegex.Extensions.RegexExtensions;
+using StreamRegex.Extensions.StringMethods;
 using StreamRegex.Lib.DFA;
 using StreamRegex.Lib.NFA;
