+++ StreamRegex.Benchmarks/LargeFileBenchmarks.cs
@@ -1,7 +1,8 @@
 ﻿using System.Text;
 using System.Text.RegularExpressions;
 using BenchmarkDotNet.Attributes;
-using StreamRegex.Extensions;
+using StreamRegex.Extensions.RegexExtensions;
+using StreamRegex.Extensions.StringMethods;
 using StreamRegex.Lib.DFA;
 using StreamRegex.Lib.NFA;
