+++ StreamRegex.Extensions/StringMethods/StringMethodExtensionsAsync.cs
@@ -1,14 +1,16 @@
 ﻿using System;
 using System.IO;
+using System.Threading.Tasks;
+using StreamRegex.Extensions.Core;
 
-namespace StreamRegex.Extensions;
+namespace StreamRegex.Extensions.StringMethods;
 
 /// <summary>
-/// Synchronous methods to emulate <see cref="String.Contains(string)"/> and <see cref="String.IndexOf(string)"/> operations on Streams.
+/// Asynchronous methods to emulate <see cref="string.Contains(string)"/> and <see cref="string.IndexOf(string)"/> operations on Streams.
 /// </summary>
-public static class StringMethodExtensions
+public static class StringMethodExtensionsAsync
 {
-    
+
     /// <summary>
     /// Check if a given string is contained in the given <see cref="Stream"/>.
     /// Using a <see cref="System.Text.RegularExpressions.Regex"/> may be faster.
@@ -18,9 +20,9 @@ public static class StringMethodExtensions
     /// <param name="comparisonType"><see cref="StringComparison"/> type to use</param>
     /// <param name="options">The <see cref="SlidingBufferOptions"/> specifying the sizes of the buffer and the overlap.</param>
     /// <returns>True if <paramref name="value"/> is contained in <paramref name="streamToCheck"/></returns>
-    public static bool Contains(this Stream streamToCheck, string value, StringComparison? comparisonType = null, SlidingBufferOptions? options = null)
+    public static async Task<bool> ContainsAsync(this Stream streamToCheck, string value, StringComparison? comparisonType = null, SlidingBufferOptions? options = null)
     {
-        return new StreamReader(streamToCheck).Contains(value, comparisonType, options);
+        return await new StreamReader(streamToCheck).ContainsAsync(value, comparisonType, options);
     }
 
     /// <summary>
@@ -32,11 +34,11 @@ public static class StringMethodExtensions
     /// <param name="comparisonType"><see cref="StringComparison"/> type to use</param>
     /// <param name="options">The <see cref="SlidingBufferOptions"/> specifying the sizes of the buffer and the overlap.</param>
     /// <returns>True if <paramref name="value"/> is contained in <paramref name="streamReaderToCheck"/></returns>
-    public static bool Contains(this StreamReader streamReaderToCheck, string value, StringComparison? comparisonType = null, SlidingBufferOptions? options = null)
+    public static async Task<bool> ContainsAsync(this StreamReader streamReaderToCheck, string value, StringComparison? comparisonType = null, SlidingBufferOptions? options = null)
     {
-        return comparisonType is { } notNullComparison ?
-            streamReaderToCheck.IsMatch(contentChunk => contentChunk.Contains(value, notNullComparison), options) :
-            streamReaderToCheck.IsMatch(contentChunk => contentChunk.Contains(value, StringComparison.Ordinal), options);
+        return await (comparisonType is { } notNullComparison ?
+            streamReaderToCheck.IsMatchAsync(contentChunk => contentChunk.Contains(value, notNullComparison), options) :
+            streamReaderToCheck.IsMatchAsync(contentChunk => contentChunk.Contains(value), options));
     }
 
     /// <summary>
@@ -48,9 +50,10 @@ public static class StringMethodExtensions
     /// <param name="comparisonType"><see cref="StringComparison"/> type to use</param>
     /// <param name="options">The <see cref="SlidingBufferOptions"/> specifying the sizes of the buffer and the overlap.</param>
     /// <returns>The index offset relative to the position when <paramref name="streamToCheck"/> was provided or -1 if not found.</returns>
-    public static long IndexOf(this Stream streamToCheck, string value, StringComparison comparisonType = StringComparison.CurrentCulture, SlidingBufferOptions? options = null)
+
+    public static async Task<long> IndexOfAsync(this Stream streamToCheck, string value, StringComparison? comparisonType = null, SlidingBufferOptions? options = null)
     {
-        return new StreamReader(streamToCheck).IndexOf(value, comparisonType, options);
+        return await new StreamReader(streamToCheck).IndexOfAsync(value, comparisonType, options);
     }
 
     /// <summary>
@@ -62,7 +65,8 @@ public static class StringMethodExtensions
     /// <param name="comparisonType"><see cref="StringComparison"/> type to use</param>
     /// <param name="options">The <see cref="SlidingBufferOptions"/> specifying the sizes of the buffer and the overlap.</param>
     /// <returns>The index offset relative to the position when <paramref name="streamReaderToCheck"/> was provided or -1 if not found.</returns>
-    public static long IndexOf(this StreamReader streamReaderToCheck, string value, StringComparison comparisonType = StringComparison.CurrentCulture, SlidingBufferOptions? options = null)
+
+    public static async Task<long> IndexOfAsync(this StreamReader streamReaderToCheck, string value, StringComparison? comparisonType = null, SlidingBufferOptions? options = null)
     {
         var opts = options ?? new SlidingBufferOptions();
         // Ensure that we can find the target string
@@ -70,15 +74,15 @@ public static class StringMethodExtensions
         {
             opts.OverlapSize = value.Length;
         }
-        var match = streamReaderToCheck.GetFirstMatch(contentChunk =>
+        var match = await streamReaderToCheck.GetFirstMatchAsync(contentChunk =>
         {
-            var idx = contentChunk.IndexOf(value, comparisonType);
+            var idx = comparisonType is { } notNullComparison ? contentChunk.IndexOf(value, notNullComparison) : contentChunk.IndexOf(value);
             if (idx != -1)
             {
-                return new SlidingBufferValueMatch(true, idx, idx+value.Length);
+                return new SlidingBufferMatch(true, idx, contentChunk[idx..(idx + value.Length)]);
             }
 
-            return new SlidingBufferValueMatch();
+            return new SlidingBufferMatch();
         }, options);
         return match.Index;
     }
